Refactor writing of string and location definitions
Add classes for writing string and location definitions based on
maps used to label their values as they are read from the trace.

Ensure these classes are closed only when the connection is closed
to ensure that all definitions to be written have been added to
the internal maps and have had IDs generated.